speech2text/server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
numpy>=1.25.0
openai-whisper==20231117
torch>=2.2.0
tokenizers>=0.13.3 --only-binary tokenizers
transformers>=4.33.2
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1
python-dotenv>=1.0.1
websockets==11.0.3
websocket-client==1.6.1 